+++ packages/frontend/src/utils/ThumbnailGenerationUtil.js
